+++ compose.yml
@@ -22,7 +22,6 @@ x-service-template: &service-template
     - node_modules_ghost_members-csv:/home/ghost/ghost/members-csv/node_modules:delegated
     - node_modules_ghost_mw-error-handler:/home/ghost/ghost/mw-error-handler/node_modules:delegated
     - node_modules_ghost_mw-vhost:/home/ghost/ghost/mw-vhost/node_modules:delegated
-    - node_modules_ghost_offers:/home/ghost/ghost/offers/node_modules:delegated
     - node_modules_ghost_post-events:/home/ghost/ghost/post-events/node_modules:delegated
     - node_modules_ghost_post-revisions:/home/ghost/ghost/post-revisions/node_modules:delegated
     - node_modules_ghost_prometheus-metrics:/home/ghost/ghost/prometheus-metrics/node_modules:delegated
@@ -189,7 +188,6 @@ volumes:
   node_modules_ghost_members-csv: {}
   node_modules_ghost_mw-error-handler: {}
   node_modules_ghost_mw-vhost: {}
-  node_modules_ghost_offers: {}
   node_modules_ghost_post-events: {}
   node_modules_ghost_post-revisions: {}
   node_modules_ghost_prometheus-metrics: {}
