+++ ghost/core/core/server/services/offers/domain/events/OfferCreatedEvent.js
