+++ ghost/core/core/server/services/offers/domain/models/OfferDescription.js
