+++ ghost/core/core/server/services/offers/domain/models/OfferTitle.js
