+++ ghost/core/core/server/services/offers/domain/models/shared/ValueObject.js
