+++ ghost/core/core/server/services/offers/service.js
@@ -1,10 +1,9 @@
-const OffersModule = require('@tryghost/members-offers');
-
 const DynamicRedirectManager = require('../lib/DynamicRedirectManager');
 const config = require('../../../shared/config');
 const urlUtils = require('../../../shared/url-utils');
 const models = require('../../models');
 const OfferBookshelfRepository = require('./OfferBookshelfRepository');
+const OffersModule = require('./OffersModule');
 
 let redirectManager;
