+++ ghost/core/package.json
@@ -100,7 +100,6 @@
     "@tryghost/limit-service": "1.2.15",
     "@tryghost/logging": "2.4.21",
     "@tryghost/members-csv": "0.0.0",
-    "@tryghost/members-offers": "0.0.0",
     "@tryghost/metrics": "1.0.37",
     "@tryghost/mw-error-handler": "0.0.0",
     "@tryghost/mw-vhost": "0.0.0",
