+++ ghost/core/test/unit/server/services/offers/domain/models/OfferDescription.test.js
@@ -1,4 +1,6 @@
-const OfferDescription = require('../../../../lib/domain/models/OfferDescription');
+const should = require('should');
+
+const OfferDescription = require('../../../../../../../core/server/services/offers/domain/models/OfferDescription');
 
 describe('OfferDescription', function () {
     describe('OfferDescription.create factory', function () {
